+++ src/main/java/janggi/dao/GameInformationDao.java
@@ -1,11 +1,8 @@
 package janggi.dao;
 
 import janggi.game.GameInformation;
-import janggi.game.MovePieceCommand;
-import janggi.setting.CampType;
 import janggi.setting.GameState;
 import janggi.setting.PieceAssignType;
-import janggi.value.Position;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.PreparedStatement;
@@ -16,7 +13,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class GameDao {
+public class GameInformationDao {
 
     private static final String SERVER = "localhost:13306";
     private static final String DATABASE = "janggi";
@@ -82,48 +79,6 @@ public class GameDao {
             preparedStatement.setString(1, GameState.END.toString());
             preparedStatement.setInt(2, gameId);
             preparedStatement.executeUpdate();
-        } catch (final SQLException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public void addMovePieceCommand(int gameId, MovePieceCommand movePieceCommand) {
-        String query =
-                "INSERT INTO movePieceRecords "
-                        + "(gameId, campType, targetPieceXPostion, targetPieceYPostion, destinationXPostion, destinationYPostion) "
-                        + "VALUES(?, ?, ?, ?, ?, ?)";
-        try {
-            Connection connection = getConnection();
-            PreparedStatement preparedStatement = connection.prepareStatement(query);
-            preparedStatement.setInt(1, gameId);
-            preparedStatement.setString(2, movePieceCommand.getCampType().toString());
-            preparedStatement.setInt(3, movePieceCommand.getTargetPiecePosition().x());
-            preparedStatement.setInt(4, movePieceCommand.getTargetPiecePosition().y());
-            preparedStatement.setInt(5, movePieceCommand.getDestination().x());
-            preparedStatement.setInt(6, movePieceCommand.getDestination().y());
-            preparedStatement.executeUpdate();
-        } catch (final SQLException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public List<MovePieceCommand> finaAllMovePieceCommand(int gameId) {
-        String query = "select * from movePieceRecords WHERE movePieceRecords.gameId = ? ORDER BY movePieceRecords.created_at ASC;";
-        try {
-            Connection connection = getConnection();
-            PreparedStatement preparedStatement = connection.prepareStatement(query);
-            preparedStatement.setInt(1, gameId);
-            ResultSet resultSet = preparedStatement.executeQuery();
-            List<MovePieceCommand> queryResult = new ArrayList<>();
-            while (resultSet.next()) {
-                MovePieceCommand command = new MovePieceCommand(
-                        CampType.valueOf(resultSet.getString("campType")),
-                        new Position(resultSet.getInt("targetPieceXPostion"), resultSet.getInt("targetPieceYPostion")),
-                        new Position(resultSet.getInt("destinationXPostion"), resultSet.getInt("destinationYPostion"))
-                );
-                queryResult.add(command);
-            }
-            return Collections.unmodifiableList(queryResult);
         } catch (final SQLException e) {
             throw new RuntimeException(e);
         }
